copyPiece, deletePiece, rotatePiece combined in piece.c
