+++ Makefile
@@ -7,6 +7,7 @@ DIR_SRC = src
 SRCS =	$(DIR_SRC)/tetris.c	\
 		$(DIR_SRC)/timer.c	\
 		$(DIR_SRC)/score.c	\
+		$(DIR_SRC)/piece.c	\
 		$(DIR_SRC)/piece_template.c	\
 		$(DIR_SRC)/view.c	\
