+++ src/tetris.c
@@ -5,33 +5,6 @@ char GameOn = TRUE;
 
 Piece current;
 
-Piece copyPiece(Piece piece)
-{
-	Piece copied = piece;
-	char **array = piece.array;
-	copied.array = (char **)malloc(copied.width * sizeof(char *));
-	int i, j;
-	for (i = 0; i < copied.width; i++)
-	{
-		copied.array[i] = (char *)malloc(copied.width * sizeof(char));
-		for (j = 0; j < copied.width; j++)
-		{
-			copied.array[i][j] = array[i][j];
-		}
-	}
-	return copied;
-}
-
-void deletePiece(Piece piece)
-{
-	int i;
-	for (i = 0; i < piece.width; i++)
-	{
-		free(piece.array[i]);
-	}
-	free(piece.array);
-}
-
 int checkPosition(Piece piece) // check position of copied piece
 {
 	char **array = piece.array;
@@ -54,7 +27,7 @@ int checkPosition(Piece piece) // check position of copied piece
 
 void getNewPiece() // return s rondom piece
 {
-	Piece new = copyPiece(randomPieces());
+	Piece new = copyPiece(getRandomPiece());
 	new.col = rand() % (COLS - new.width + 1);
 	new.row = 0;
 	deletePiece(current);
@@ -63,21 +36,6 @@ void getNewPiece() // return s rondom piece
 	{
 		GameOn = FALSE;
 	}
-}
-
-void rotatePiece(Piece piece) // rotates clockwise
-{
-	Piece tmp = copyPiece(piece);
-	int i, j, k, width;
-	width = piece.width;
-	for (i = 0; i < width; i++)
-	{
-		for (j = 0, k = width - 1; j < width; j++, k--)
-		{
-			piece.array[i][j] = tmp.array[k][i];
-		}
-	}
-	deletePiece(tmp);
 }
 
 void writeToTable()
@@ -117,8 +75,6 @@ void checkLines()
 	}
 	addScore(count);
 }
-
-
 
 void manipulateCurrent(int action)
 {
